Fixed the indenting of the when condition
Tried fixing this issue for around a hour until I asked someone that saw this in a second. It was basically setting the value of `“ansible_distribution_release == "xenial”` and `ansible_distribution_release != "xenial”` to the variable named `when` instead of being an condition for the variable `mysql_conf_path`.